refactor(sync): separate namespace reconciliation
