+++ src/sync.rs
@@ -31,8 +31,7 @@ use crate::model::{
     CommitRecord, Cursor, Manifest, NamespaceChange, Node, NodeKind, OperationId, VolumeId,
 };
 use crate::replica::{
-    CommonBase, Conflict, ConflictKind, PendingMaterialization, PendingPublication, ReplicaLock,
-    ReplicaPaths, ReplicaState,
+    CommonBase, PendingMaterialization, PendingPublication, ReplicaLock, ReplicaPaths, ReplicaState,
 };
 use crate::store::{DataStore, MetadataStore, PublicationOutcome};
 
@@ -147,7 +146,7 @@ pub(crate) async fn sync_once(volume: &ManagedVolume, request: SyncRequest<'_>) 
     {
         return Ok(observed.head.cursor.generation);
     }
-    let (target, conflicts) = merge(
+    let (target, conflicts) = crate::reconcile::merge(
         &base,
         &stable_local,
         &remote,
@@ -175,7 +174,7 @@ pub(crate) async fn sync_once(volume: &ManagedVolume, request: SyncRequest<'_>) 
         return Ok(state.common.as_ref().unwrap().cursor.generation);
     }
 
-    let changes = diff(&remote, &target);
+    let changes = crate::reconcile::diff(&remote, &target);
     let operation = OperationId::parse(Uuid::new_v4().to_string())?;
     let pending = PendingPublication {
         operation: operation.clone(),
@@ -359,163 +358,6 @@ async fn remote_manifest(
         bail!("Managed change log did not reach its fixed target cursor");
     }
     Ok(manifest)
-}
-
-fn diff(parent: &Manifest, target: &Manifest) -> Vec<NamespaceChange> {
-    let mut changes = Vec::new();
-    for (path, node) in &parent.entries {
-        if !target.entries.contains_key(path) {
-            changes.push(NamespaceChange::Remove {
-                path: path.clone(),
-                removed: node.id.clone(),
-            });
-        }
-    }
-    for (path, node) in &target.entries {
-        let previous = parent.entries.get(path);
-        if previous != Some(node) {
-            changes.push(NamespaceChange::Put {
-                path: path.clone(),
-                node: node.clone(),
-                replaces: previous.map(|value| value.id.clone()),
-            });
-        }
-    }
-    changes
-}
-
-fn merge(
-    base: &Manifest,
-    local: &Manifest,
-    remote: &Manifest,
-    cursor: &Cursor,
-    resolved: &BTreeSet<String>,
-) -> (Manifest, Vec<Conflict>) {
-    let base_identities = identities(base);
-    let local_identities = identities(local);
-    let remote_identities = identities(remote);
-    let mut identity_choices = BTreeMap::<String, Option<Node>>::new();
-    let mut conflicts = Vec::new();
-    for (id, (base_path, base_node)) in &base_identities {
-        let local_value = local_identities.get(id);
-        let remote_value = remote_identities.get(id);
-        let local_path = local_value.map(|(path, _)| path.as_str());
-        let remote_path = remote_value.map(|(path, _)| path.as_str());
-        let local_relocated = local_path != Some(base_path.as_str());
-        let remote_relocated = remote_path != Some(base_path.as_str());
-        let divergent_location = local_relocated && remote_relocated && local_path != remote_path;
-        let rename_vs_edit = local_path.is_some()
-            && local_relocated
-            && remote_path == Some(base_path.as_str())
-            && remote_value.is_some_and(|(_, node)| node != base_node)
-            || remote_path.is_some()
-                && remote_relocated
-                && local_path == Some(base_path.as_str())
-                && local_value.is_some_and(|(_, node)| node != base_node);
-        if !divergent_location && !rename_vs_edit {
-            continue;
-        }
-        for path in [Some(base_path.as_str()), local_path, remote_path]
-            .into_iter()
-            .flatten()
-        {
-            identity_choices.insert(path.to_owned(), local.entries.get(path).cloned());
-        }
-        if !resolved.contains(base_path) {
-            conflicts.push(Conflict {
-                path: base_path.clone(),
-                kind: ConflictKind::DivergentRename,
-                base: Some(base_node.clone()),
-                local: local_value.map(|(_, node)| node.clone()),
-                remote: remote_value.map(|(_, node)| node.clone()),
-                remote_cursor: cursor.clone(),
-            });
-        }
-    }
-    let paths = base
-        .entries
-        .keys()
-        .chain(local.entries.keys())
-        .chain(remote.entries.keys())
-        .cloned()
-        .collect::<BTreeSet<_>>();
-    let mut entries = BTreeMap::new();
-    for path in paths {
-        let base_node = base.entries.get(&path);
-        let local_node = local.entries.get(&path);
-        let remote_node = remote.entries.get(&path);
-        let selected = if let Some(node) = identity_choices.get(&path) {
-            node.as_ref()
-        } else if resolved.contains(&path) {
-            local_node
-        } else if local_node == base_node {
-            remote_node
-        } else if remote_node == base_node || local_node == remote_node {
-            local_node
-        } else if independently_created_directories(
-            base_node,
-            local_node,
-            remote_node,
-            &base_identities,
-        ) {
-            // Directory identities are internal bookkeeping. When both replicas
-            // created the same previously absent path, retain the authority's
-            // identity and merge their children path by path.
-            remote_node
-        } else {
-            conflicts.push(Conflict {
-                path: path.clone(),
-                kind: conflict_kind(base_node, local_node, remote_node),
-                base: base_node.cloned(),
-                local: local_node.cloned(),
-                remote: remote_node.cloned(),
-                remote_cursor: cursor.clone(),
-            });
-            local_node
-        };
-        if let Some(node) = selected {
-            entries.insert(path, node.clone());
-        }
-    }
-    (Manifest { entries }, conflicts)
-}
-
-fn independently_created_directories(
-    base: Option<&Node>,
-    local: Option<&Node>,
-    remote: Option<&Node>,
-    base_identities: &BTreeMap<crate::model::NodeId, (String, Node)>,
-) -> bool {
-    let (Some(local), Some(remote)) = (local, remote) else {
-        return false;
-    };
-    base.is_none()
-        && matches!(local.kind, NodeKind::Directory)
-        && matches!(remote.kind, NodeKind::Directory)
-        && local.id != remote.id
-        && !base_identities.contains_key(&local.id)
-        && !base_identities.contains_key(&remote.id)
-}
-
-fn identities(manifest: &Manifest) -> BTreeMap<crate::model::NodeId, (String, Node)> {
-    manifest
-        .entries
-        .iter()
-        .map(|(path, node)| (node.id.clone(), (path.clone(), node.clone())))
-        .collect()
-}
-
-fn conflict_kind(base: Option<&Node>, local: Option<&Node>, remote: Option<&Node>) -> ConflictKind {
-    match (base, local, remote) {
-        (Some(_), None, Some(_)) | (Some(_), Some(_), None) => ConflictKind::DeleteVsModify,
-        (_, Some(a), Some(b))
-            if std::mem::discriminant(&a.kind) != std::mem::discriminant(&b.kind) =>
-        {
-            ConflictKind::IncompatibleTypeReplacement
-        }
-        (_, Some(a), Some(b)) if a.id == b.id => ConflictKind::SameNodeModified,
-        _ => ConflictKind::DivergentRename,
-    }
 }
 
 async fn upload_changes(
@@ -825,6 +667,8 @@ fn resolution_paths(paths: &[PathBuf]) -> Result<BTreeSet<String>> {
 mod tests {
     use super::*;
     use crate::model::{ContentRef, NodeId};
+    use crate::reconcile::{diff, merge};
+    use crate::replica::ConflictKind;
 
     fn directory(id: &str) -> Node {
         Node {
